+++ sdk/azcore/CHANGELOG.md
@@ -6,14 +6,13 @@
 * Removed `arm.Connection`
 * Removed `azcore.Credential` and `.NewAnonymousCredential()`
   * `NewRPRegistrationPolicy` now requires an `azcore.TokenCredential`
-* Removed `runtime.NewPipeline`
+* `runtime.NewPipeline` has a new signature that simplifies implementing custom authentication
 * `arm/runtime.RegistrationOptions` embeds `policy.ClientOptions`
 
 ### Features Added
 * Updating Documentation
 * Added string typdef `arm.Endpoint` to provide a hint toward expected ARM client endpoints
 * `azcore.ClientOptions` contains common pipeline configuration settings
-* Added `runtime.NewDefaultPipeline`, a constructor of default data plane pipelines
 
 ### Bug Fixes
 * Fixed a potential panic when creating the default Transporter.